refactor(cli): move logs tests to test_cli_logs.py, keep test_cli_commands.py under 1000 lines
